# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Work In Progress",
    "description": "Dokumentasi API untuk proses inspeksi, SWA, dan validasi",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/inspection/template": {
      "get": {
        "tags": ["Inspection"],
        "summary": "Get inspection template data by ID and optional isDraft",
        "operationId": "viewInspectionTemplate",
        "parameters": [
          {
            "name": "IsDraft",
            "in": "query",
            "description": "Filter by active isDraft (true/false)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Template data retrieved successfully"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Template not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": ["Inspection"],
        "summary": "Create a new inspection template",
        "operationId": "ddc67411221a4d27718545e8f7b6e22d",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "Content",
                  "Description",
                  "IsActivated",
                  "CreatedBy"
                ],
                "properties": {
                  "Content": {
                    "type": "object",
                    "example": {
                      "field": "value"
                    }
                  },
                  "Description": {
                    "type": "string",
                    "example": "Template Deskripsi"
                  },
                  "IsActivated": {
                    "type": "boolean",
                    "example": true
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Template created successfully"
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/inspection/template-content/{templateId}": {
      "get": {
        "tags": ["Inspection"],
        "summary": "Get full template content by ID",
        "operationId": "inspectionTemplateById",
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "description": "Template ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Template content retrieved successfully"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Template not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/inspection/template/{templateId}": {
      "put": {
        "tags": ["Inspection"],
        "summary": "Update an existing inspection template",
        "operationId": "db014741e6dd47898dd1eef747249ad3",
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "description": "Template ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "Content": {
                    "type": "object",
                    "example": {
                      "field": "value"
                    }
                  },
                  "Description": {
                    "type": "string",
                    "example": "Template Deskripsi"
                  },
                  "IsActivated": {
                    "type": "boolean",
                    "example": true
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update success",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Template updated successfully"
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["Inspection"],
        "summary": "Delete an inspection template",
        "operationId": "f8648bda45ae371f7b4daaa802bac12b",
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "description": "Template ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Template deleted successfully"
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/inspection/templates": {
      "get": {
        "tags": ["Inspection"],
        "summary": "Get all inspection templates created by current user",
        "operationId": "c5556633d88987ea1c8ae61f169c01ba",
        "parameters": [
          {
            "name": "perPage",
            "in": "query",
            "description": "Number of templates per page",
            "schema": {
              "type": "integer",
              "example": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of inspection templates",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/inspection/template/{templateId}/enabled": {
      "put": {
        "tags": ["Inspection"],
        "summary": "Enable or disable an inspection template",
        "operationId": "99d6ed4187c60502f7f75685279a8637",
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "description": "Template ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["IsActivated"],
                "properties": {
                  "IsActivated": {
                    "type": "boolean",
                    "example": true
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Template activation updated",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/inspection": {
      "post": {
        "tags": ["Inspection"],
        "summary": "Create a new inspection record",
        "operationId": "e5a511c19534d4f08026b74f66818dcb",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "SWAPIC",
                  "Location",
                  "WorkType",
                  "InspectionDate",
                  "WorkingPermit",
                  "InspectionForm",
                  "isDraft",
                  "CreatedBy"
                ],
                "properties": {
                  "SWAPIC": {
                    "type": "string",
                    "example": "SW-001"
                  },
                  "Location": {
                    "type": "string",
                    "example": "Site A"
                  },
                  "WorkType": {
                    "type": "string",
                    "example": "Electrical"
                  },
                  "InspectionDate": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-05-10T10:00:00Z"
                  },
                  "InspectionForm": {
                    "type": "object",
                    "example": {
                      "field": "value"
                    }
                  },
                  "IsDraft": {
                    "type": "boolean",
                    "example": true
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inspection created successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/inspection/{inspectionId}": {
      "get": {
        "tags": ["Inspection"],
        "summary": "Get inspection details by ID",
        "operationId": "bfde84d2b6c3a3bcb3ae22c9e7e6ecc9",
        "parameters": [
          {
            "name": "inspectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Inspection details"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": ["Inspection"],
        "summary": "Update an existing inspection",
        "operationId": "4b837ea59b248682ae11e02dccb7700e",
        "parameters": [
          {
            "name": "inspectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "SWAPIC",
                  "Location",
                  "WorkType",
                  "InspectionDate",
                  "WorkingPermit",
                  "InspectionForm",
                  "isDraft"
                ],
                "properties": {
                  "SWAPIC": {
                    "type": "string",
                    "example": "SW-001"
                  },
                  "Location": {
                    "type": "string",
                    "example": "Site A"
                  },
                  "WorkType": {
                    "type": "string",
                    "example": "Electrical"
                  },
                  "InspectionForm": {
                    "type": "object",
                    "example": {
                      "field": "value"
                    }
                  },
                  "IsDraft": {
                    "type": "boolean",
                    "example": true
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inspection updated successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["Inspection"],
        "summary": "Delete an inspection by ID",
        "operationId": "02294e5a6e7970e80a8d7ae289732327",
        "parameters": [
          {
            "name": "inspectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Inspection deleted"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/inspection/by-user": {
      "get": {
        "tags": ["Inspection"],
        "summary": "View inspection list based on authenticated user's role",
        "operationId": "e0083835d874ad4099ac276f465b0510",
        "parameters": [
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 10
            }
          },
          {
            "name": "IsDraft",
            "in": "query",
            "description": "Filter by active isDraft (true/false)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Inspections retrieved"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/inspection/chart": {
      "get": {
        "tags": ["Inspection"],
        "summary": "Get inspection chart grouped by time period",
        "description": "Return inspection count grouped by day, week, month, or year for authenticated user",
        "operationId": "a9282dc6a134d75c1ec7fc4efceebf1c",
        "parameters": [
          {
            "name": "group_by",
            "in": "query",
            "description": "Group chart data by time period",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["day", "week", "month", "year"],
              "example": "day"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Inspections retrieved"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Use a bearer token to access this API",
        "name": "Authorization",
        "in": "header",
        "bearerFormat": "JWT",
        "scheme": "bearer"
      }
    }
  },
...
}